Use correct architectures value in library.properties
When an ESP8266 board is selected the previous incorrectly capitalized architectures value ESP8266 causes the RTC4543lib example sketch to not appear in the File > Examples menu and the warning:

WARNING: library RTC4543lib-master claims to run on (avr, tiny, ESP8266) architecture(s) and may be incompatible with your current board which runs on (esp8266) architecture(s).

is shown during compilation.